[change] sort_environ()変更
sort_environ()内の変数についてsize_t型を使用するように変更。

+++ srcs/utils/print_sorted_env.c
@@ -51,7 +51,7 @@ static void	merge(char **a, char **b, size_t index[3])
 
 void		sort_environ(char **a, char **b, size_t front, size_t end)
 {
-	int mid;
+	size_t mid;
 	size_t index[3];
 
 	if (front == end || front+ 1 == end)
